requirements.txt
alpaca==1.0.0
alpaca-py==0.7.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
attrs==22.2.0
better-exceptions==0.3.3
bleach==6.0.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.95.0
Flask==2.2.3
fonttools==4.39.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
icecream==2.1.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
msgpack==1.0.4
newtulipy==0.4.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
param==1.13.0
Pillow==9.4.0
plotly==5.13.1
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyct==0.5.0
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pysos==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyYAML==6.0
requests==2.28.2
rich==13.3.1
rsa==4.9
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sseclient-py==1.7.2
starlette==0.26.1
streamlit==1.20.0
structlog==23.1.0
tenacity==8.2.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.14
uvicorn==0.21.1
validators==0.20.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
zipp==3.15.0
